+++ c_api/IndexIVF_c.cpp
@@ -80,12 +80,12 @@ int faiss_IndexIVF_make_direct_map(FaissIndexIVF* index,
 }
 
 double faiss_IndexIVF_imbalance_factor (const FaissIndexIVF* index) {
-    return reinterpret_cast<const IndexIVF*>(index)->imbalance_factor();
+    return reinterpret_cast<const IndexIVF*>(index)->invlists->imbalance_factor();
 }
 
 /// display some stats about the inverted lists
 void faiss_IndexIVF_print_stats (const FaissIndexIVF* index) {
-    reinterpret_cast<const IndexIVF*>(index)->print_stats();
+    reinterpret_cast<const IndexIVF*>(index)->invlists->print_stats();
 }
 
 void faiss_IndexIVFStats_reset(FaissIndexIVFStats* stats) {
