+++ c_api/MetaIndexes_c.h
@@ -44,34 +44,6 @@ int faiss_IndexIDMap2_new(FaissIndexIDMap2** p_index, FaissIndex* index);
 /// make the rev_map from scratch
 int faiss_IndexIDMap2_construct_rev_map(FaissIndexIDMap2* index);
 
-/** Index that concatenates the results from several sub-indexes
- */
-FAISS_DECLARE_CLASS_INHERITED(IndexShards, Index)
-
-FAISS_DECLARE_GETTER_SETTER(IndexShards, int, own_fields)
-FAISS_DECLARE_GETTER_SETTER(IndexShards, int, threaded)
-FAISS_DECLARE_GETTER_SETTER(IndexShards, int, successive_ids)
-
-int faiss_IndexShards_new(FaissIndexShards** p_index, idx_t d);
-
-int faiss_IndexShards_new_with_options(FaissIndexShards** p_index, idx_t d, int threaded, int successive_ids);
-
-/** get a pointer to the index' shards (the `shard_indexes` field). The
- * outputs of this function become invalid after any operation that can modify the index.
- * 
- * @param index   opaque pointer to index object
- * @param p_shard_indexes    output, the pointer to the beginning of `shard_indexes`.
- * @param p_size  output, the current length of `shard_indexes`.
- */
-void faiss_IndexShards_shard_indexes(FaissIndexShards* index, FaissIndex** p_shard_indexes, size_t* p_size);
-
-int faiss_IndexShards_add_shard(FaissIndexShards* index, FaissIndex* shard);
-
-/// update metric_type and ntotal
-int faiss_IndexShards_sync_with_shard_indexes(FaissIndexShards* index);
-
-FaissIndex* faiss_IndexShards_at(FaissIndexShards* index, int i);
-
 #ifdef __cplusplus
 }
 #endif
